tests: move the server tests to a separate file
